Remove version command from emacs and vi mode.
The command isn't bound to a key per default and the functionality (and
implementation) seems rather odd. The keybinding is strictly for
interactive use, so I don't expect the removal to break any scripts. The
KSH_VERSION environment variable is still available for interactive and
non-interactive use.

no objection oga@ pedro@