fdt: recursively translate memory addresses
Before this commit we only translated a devices' memory information
once, using the information of its parent bus. This is not good, as
that parent bus can also be part of a translated bus. This means we
need to go all the way up and translate the addresses until we have
addresses that the CPU can use.

If we hit an empty range property, this means we hit a 1:1 mapping.
Thus we can go one level upwards and check the next range property.
If there's no range property, we hit a non-translatable barrier and
have finished our translation process.