fdt: fix memory translation via ranges
On translation, we used the wrong address- and size-cells counts.

Before this commit, we had the following scenario:
  * node's memory address needs to be translated
  * lookup node's address and size cells
  * lookup parent's address and size cells
  * ...

This doesn't work, as the node's address cells do not say how
big it's own cells are, but instead how big the ones of its
childrens are. Therefore it will change to:
  * node's memory address needs to be translated
  * lookup parent's address and size cells
  * lookup parent's parent's cells

To simplify this, we only need to change the context the
translation runs in. Previously the translation was based
on the node. Basing it on the parent instead fixes that.