fdt: attach an attached devices' devices
When a device is attached, we used to not iterate over its
children anymore. But when we have a bus controller, that
provides some API but otherwise is still a normal simple-bus,
we need to be able to also attach its children.

This might not be the best approach, but until there's a better
implementation we can use this kind of workaround.