com(4): support for the NS16750
The NS16750 has a specific interrupt that tells you it's busy.
To stop this interrupt from happening we need to read the usr
register and wait until it's ready.

To make sure this is only used on a hardware that supports this,
make the attachment driver set a hardware flag.