fdt: attach NS16550 compatible consoles via com(4)
sunxi's sxiuart is basically an NS16550 with a specialty
in the interrupt handling, Thus we can just use com(4).

Still, we need to use the correct frequencies. Unless
the clocks have been attached that early, which they
are not, we don't really know the frequency. Thus we
need to make some kind of educated guess.