arm: remove bus_space_map hook used for early peripherals
This code used to make it possible to continue using addresses
for peripherals which were mapped during bootstrap.

This means that the virtual address we got on the early bootup
could be used as physical address in a subsequent bus_space_map
call, to be able to continue using that mapping.

Shortly discussed with drahn@