Having the same settings on the same phy across multiple boards is just asking too much in the arm world.
Translate the fec parameters from the novena dtb to set a different clock
skew to the same micrel phy used on sabre lite.  The novena dtb sets
txd[0-3]-skew-ps to 3000, the sabre lite sets them to 0.  When run through
the shifting/oring process from the micrel phy driver in FreeBSD this ends
up being a write of 0xffff on the novena and 0x0000 on sabre lite when
writing to TX_DATA_PAD_SKEW.

This change resolves the stability problems djm was seeing with imxenet
on novena.

ok djm@

Imported by patrick@