fdt: fix matching compatible nodes
An "compatible" property is a list of strings, where each string
is terminated with the usual null character. If we're looking to
match a "compatible" with this property we need to go through every
single string and see how they compare. This has already been
implemented. Still, there was one issue with comparing the following:

  "arm,gic" vs. "arm,gic-v2m-frame"

Instead of making sure the string is the complete same, only the
first 7 characters where checked, which made the comparison return
"it's all good". This is not wanted.

To fix this issue, check the length of both before starting any
other string comparison.