introduce SRPs to the interface index maps to make if_get() mpsafe.
each index in the ifidx map is now an SRP that points at each
interfaces struct ifnet.

because the map needs to be grown at run time as interfaces are
added to the system, we now use an SRP to gain access to the map.

if_get enters the idx map and follows it to the ifnet pointer. if
an interface exists at that index it gets adds a refcount before
leaving the ifp SRP.

modifications to the map pointer and the interface pointers are
serialised by the kernel lock.

ok mpi@ claudio@