fdt: rework console lookup and attachment
Before this commit we found our early console by:
  * looking for a known compatible node and attach the first
  * look for the /chosen/stdout-path when used on an i.MX6 SoC

The first serial in the device tree is not necessarily the actual
serial wired to the outside. Instead of doing that, we should always
check the /chosen path first, which can show the correct serial node
to be used. If that fails, we can fall back to using the first serial
node in the tree.