sunxi: dont attach sxiuart to fdt
The serial controller in the Allwinner SoCs use the Synopsys DesignWare
ABP UART. This controller is fairly compatible to the standard com(4).

For another SoC that uses the same controller a com(4) attachment
driver has been written. This driver is to be used when booted with
a device tree. Thus we don't need sxiuart on the fdt bus.