Revert SIOCDIFADDR_IN6 fix.
It's not possible to fix this issue in ifconfig(8) because in_control()
and in6_control() have a subtle semantic difference which result in
breaking alias for IPv4.

in_control() always select the first address on the list of an interface
whereas in6_control() doesn't.  That's why ifconfig(8) passes an empty
"struct in_aliasreq" and that fails in netinet6.

Breakage reported by deraadt@