Unwind a bunch of bitrig atomic/locking.
This tries to remove bitrig specific atomics and locking stuff that was
no longer maintained.  Additionally a bunch of OpenBSD stuff was pulled
in so that we can stabilize and catch up.  Unfortunately some commit
messages will be lost because of the wholesale sized hammer that was
used.

This cleanup is not complete but kernel seems to be stable.

There are still a few uvm and vfs files that must be reviewed before
this can move forward.