Fix PT_IO with PIOD_WRITE_D
PIOD_WRITE_I and PIOD_WRITE_D are supposed to be equal because we don't
implement separation of instruction and date in our ptrace
implemenatation. Because of a refactoring actually only PIOD_WRITE_I
would work as intended, fix this by getting rid of the req parameter and
just use the uio rw member to decide what to do based on the type of io.