Make acpidmar (aka vtd) a boot time option.
Select hardware runs fine with vtd but some doesn't and having this as a
branch makes it very painful to use/test.

One can now boot the kernel with -c and enable acpidmar to see if vtd
works.  If this works well enough one can then use config to modify
binary kernel to boot with acpidmar enabled.  See config(8) for
instructions.