Add a poor-mans install cd target.
This is supposed to be run by hand after src AND xenocara make snap.

Example usage:
RELEASEDIR=/usr/snap/amd64/release/ XRELEASEDIR=/usr/snap/amd64/xrelease make

There is no natural location to add this to snap targets.  Also note
that at this time it does not copy the signatures to the iso.  A nice
addition would be to add some critical ports to the installer such as
bitrig-syscomp but that is a significant change to the boot media.

add all files mentioned in SHA256, try to copy signature if it exists in RELEASEDIR